requirements.txt
Django==4.0.6
django-dotenv==1.4.2
gunicorn==20.1.0
psycopg2-binary==2.9.3